+++ src/main/java/net/imglib2/analyzesegmentation/wormfit/InlierCells.java
@@ -7,8 +7,8 @@ import java.util.concurrent.atomic.AtomicInteger;
 import org.scijava.java3d.BranchGroup;
 import org.scijava.java3d.utils.geometry.Sphere;
 import org.scijava.vecmath.Color3f;
-import org.scijava.vecmath.Point3f;
-import org.scijava.vecmath.Vector3f;
+import org.scijava.vecmath.Point3d;
+import org.scijava.vecmath.Vector3d;
 
 import customnode.CustomLineMesh;
 import ij3d.Content;
@@ -22,8 +22,8 @@ public class InlierCells
 	final protected static AtomicInteger counter = new AtomicInteger();
 
 	final List< InlierCell > inliers;
-	float r0, r1;
-	final Point3f p0, p1;
+	double r0, r1;
+	final Point3d p0, p1;
 
 	public Color3f inlierCol = new Color3f( 1, 0, 0 );
 	public Color3f vectorCol = new Color3f( 1, 0, 0 );
@@ -36,7 +36,7 @@ public class InlierCells
 	Content vector;
 	BranchGroup truncatedCone;
 
-	public InlierCells( final List< InlierCell > inliers, final float r0, final float r1, final Point3f p0, final Point3f p1 )
+	public InlierCells( final List< InlierCell > inliers, final double r0, final double r1, final Point3d p0, final Point3d p1 )
 	{
 		this.inliers = inliers;
 		this.r0 = r0;
@@ -46,13 +46,13 @@ public class InlierCells
 	}
 
 	public List< InlierCell > getInlierCells() { return inliers; }
-	public float getR0() { return r0; }
-	public float getR1() { return r1; }
-	public Point3f getP0() { return p0; }
-	public Point3f getP1() { return p1; }
+	public double getR0() { return r0; }
+	public double getR1() { return r1; }
+	public Point3d getP0() { return p0; }
+	public Point3d getP1() { return p1; }
 
-	public void setR0( final float r0 ) { this.r0 = r0; }
-	public void setR1( final float r1 ) { this.r1 = r1; }
+	public void setR0( final double r0 ) { this.r0 = r0; }
+	public void setR1( final double r1 ) { this.r1 = r1; }
 
 	public void visualizeInliers( final Image3DUniverse univ, final Cells cells, final boolean colorInliers, final boolean drawLine, final boolean drawCone )
 	{
@@ -98,12 +98,12 @@ public class InlierCells
 		// draw the truncated cone
 		if ( drawCone )
 		{
-			final Vector3f v0 = new Vector3f( p1.x - p0.x, p1.y - p0.y, p1.z - p0.z );
+			final Vector3d v0 = new Vector3d( p1.x - p0.x, p1.y - p0.y, p1.z - p0.z );
 	
-			final Point3f pc0 = new Point3f( 0, -v0.length()/2, 0 );
-			final Point3f pc1 = new Point3f( 0, v0.length()/2, 0 );
+			final Point3d pc0 = new Point3d( 0, -v0.length()/2, 0 );
+			final Point3d pc1 = new Point3d( 0, v0.length()/2, 0 );
 	
-			this.truncatedCone = Java3DHelpers.drawTruncatedCone( r0, r1, v0.length(), univ, Algebra.getTransformation( pc0, pc1, p0, p1, false ), truncatedConeColor, truncatedConeTransparency );
+			this.truncatedCone = Java3DHelpers.drawTruncatedCone( (float)r0, (float)r1, (float)v0.length(), univ, Algebra.getTransformation( pc0, pc1, p0, p1, false ), truncatedConeColor, truncatedConeTransparency );
 		}
 	}
