+++ src/main/java/net/imglib2/drift/fit/LinkedPoint.java
@@ -18,8 +18,8 @@
  */
 package net.imglib2.drift.fit;
 
-import mpicbg.imglib.util.Util;
 import mpicbg.models.Point;
+import net.imglib2.util.Util;
 
 public class LinkedPoint<P> extends Point
 {
@@ -27,13 +27,13 @@ public class LinkedPoint<P> extends Point
 
 	final P link;
 	
-	public LinkedPoint( final float[] l, final P link )
+	public LinkedPoint( final double[] l, final P link )
 	{
 		super( l.clone() );		
 		this.link = link;
 	}
 
-	public LinkedPoint( final float[] l, final float[] w, final P link )
+	public LinkedPoint( final double[] l, final double[] w, final P link )
 	{
 		super( l.clone(), w.clone() );
 		this.link = link;
